@@ -1,7 +1,7 @@
 {
   "name": "karaokemugen-app",
-  "version": "8.0.5-master",
-  "versionName": "in memory of Shami Sauvage",
+  "version": "8.0.8-master",
+  "versionName": "Sakuya Splendide",
   "description": "Karaoke playlist manager / player",
   "bin": {
     "karaokemugen": "./dist/src/index.js"
@@ -83,7 +83,8 @@
   "license": "MIT",
   "private": true,
   "dependencies": {
-    "@sentry/electron": "^5.0.0",
+    "@hiogawa/flac-picture": "^1.0.0-pre.5",
+    "@sentry/electron": "^5.1.0",
     "agentkeepalive": "^4.5.0",
     "ass-compiler": "^0.1.11",
     "ass-parser": "^0.2.0",
@@ -100,8 +101,8 @@
     "dayjs": "^1.11.11",
     "discord-rpc": "^4.0.1",
     "dotenv": "^16.4.5",
-    "electron-updater": "^6.1.8",
-    "execa": "^9.1.0",
+    "electron-updater": "^6.2.1",
+    "execa": "^9.2.0",
     "express": "^4.19.2",
     "extract-zip": "^2.0.1",
     "fastq": "^1.17.1",
@@ -126,7 +127,7 @@
     "os-locale": "^6.0.2",
     "p-map": "^7.0.2",
     "p-retry": "^6.2.0",
-    "pg": "^8.11.5",
+    "pg": "^8.12.0",
     "pg-copy-streams": "^6.0.6",
     "portfinder": "1.0.32",
     "postgrator": "^7.2.0",
@@ -136,18 +137,18 @@
     "rxjs": "^7.8.1",
     "sanitize-filename": "^1.6.3",
     "semver": "^7.6.2",
-    "simple-git": "3.24.0",
+    "simple-git": "3.25.0",
     "slug": "^9.1.0",
     "socket.io": "4.7.5",
     "socket.io-client": "4.7.5",
     "source-map-support": "^0.5.21",
     "subsrt-ts": "^2.1.2",
-    "systeminformation": "^5.22.10",
+    "systeminformation": "^5.22.11",
     "tasklist": "^5.0.0",
     "tmi.js": "^1.8.5",
     "topological-sort": "^0.3.0",
     "ultrastar2ass": "^1.0.24",
-    "uuid": "^9.0.1",
+    "uuid": "^10.0.0",
     "validate.js": "^0.13.1",
     "which": "^4.0.0",
     "winston": "^3.13.0",
@@ -173,10 +174,10 @@
     "@types/js-yaml": "^4.0.9",
     "@types/jsonwebtoken": "^9.0.6",
     "@types/langs": "^2.0.5",
-    "@types/lodash": "4.17.4",
+    "@types/lodash": "4.17.5",
     "@types/mocha": "^10.0.6",
     "@types/multer": "^1.4.11",
-    "@types/node": "20.13.0",
+    "@types/node": "20.14.5",
     "@types/pg": "8.11.6",
     "@types/pg-copy-streams": "^1.2.5",
     "@types/qrcode": "^1.5.5",
@@ -188,26 +189,26 @@
     "@types/which": "^3.0.4",
     "@types/yesql": "^4.1.4",
     "chai": "^5.1.1",
-    "electron": "^30.0.9",
+    "electron": "^31.0.1",
     "electron-builder": "^24.13.3",
-    "esbuild": "^0.21.4",
-    "eslint": "^9.4.0",
+    "esbuild": "^0.21.5",
+    "eslint": "^9.5.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-security": "^3.0.0",
+    "eslint-plugin-security": "^3.0.1",
     "eslint-plugin-simple-import-sort": "^12.1.0",
-    "globals": "^15.3.0",
+    "globals": "^15.6.0",
     "husky": "^9.0.11",
-    "i18next-scanner": "^4.4.0",
-    "lint-staged": "^15.2.5",
+    "i18next-scanner": "^4.5.0",
+    "lint-staged": "^15.2.7",
     "mocha": "^10.4.0",
     "postgrator-cli": "^8.0.0",
-    "prettier": "3.2.5",
+    "prettier": "3.3.2",
     "rimraf": "^5.0.7",
     "ts-node": "^10.9.2",
     "typescript": "5.4.5",
-    "typescript-eslint": "^7.11.0",
+    "typescript-eslint": "^7.13.1",
     "vitest": "^1.6.0"
   },
   "optionalDependencies": {
