@@ -5,17 +5,19 @@ import './utils/isoLanguages';
 import './utils/polyfills';
 import './utils/socket';
 
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 
 import App from './App';
 import GlobalStateProvider from './store/GlobalStateProvider';
 
-ReactDOM.render(
+const container = document.getElementById('mountpoint');
+const root = createRoot(container);
+
+root.render(
 	<GlobalStateProvider>
 		<BrowserRouter>
 			<App />
 		</BrowserRouter>
-	</GlobalStateProvider>,
-	document.getElementById('mountpoint')
+	</GlobalStateProvider>
 );
