@@ -19,14 +19,14 @@ export const urls = {
 	database: () => openURL('/system/db'),
 	helpGuide: () => shell.openExternal('https://docs.karaokes.moe/user-guide/getting-started/'),
 	website: () => shell.openExternal('https://karaokes.moe'),
-	twitter: () => shell.openExternal('https://twitter.com/KaraokeMugen'),
+	twitter: () => shell.openExternal('https://x.com/KaraokeMugen'),
 	mastodon: () => shell.openExternal('https://shelter.moe/@KaraokeMugen'),
 	bluesky: () => shell.openExternal('https://bsky.app/profile/karaokes.moe'),
 	discord: () => shell.openExternal('https://karaokes.moe/discord'),
 	forum: () => shell.openExternal('https://discourse.karaokes.moe'),
 	gitlab: () => shell.openExternal('https://gitlab.com/karaokemugen/code/karaokemugen-app'),
 	changelog: () => shell.openExternal('https://gitlab.com/karaokemugen/code/karaokemugen-app/-/releases'),
-	contribDoc: () => shell.openExternal('https://docs.karaokes.moe/contrib-guide/base/'),
+	contribDoc: () => shell.openExternal('https://gitlab.com/karaokemugen/bases/karaokebase/-/wikis/home'),
 	sendKaraoke: () => shell.openExternal('https://kara.moe/base/import'),
 	reportBug: () => shell.openExternal('https://discourse.karaokes.moe/c/help/8'),
 	translations: () => shell.openExternal('https://hosted.weblate.org/projects/karaoke-mugen/'),
